requirements.txt
langchain
langchain_community
langchain-groq
python-dotenv
srt
faster-whisper
chromadb
fastapi